renamed type file to core.ts
