+++ src/views/ChartsView.vue
@@ -3,7 +3,7 @@ import { ref, type Ref } from 'vue'
 import { Duration, Icon } from '@/types/general'
 import { AppName } from '@/constants/global'
 import { useMeta } from 'quasar'
-import { recordGroups } from '@/types/database'
+import { recordGroups } from '@/types/core'
 import useUIStore from '@/stores/ui'
 import useRoutables from '@/composables/useRoutables'
 import ResponsivePage from '@/components/ResponsivePage.vue'
