+++ src/views/DataLogsView.vue
@@ -6,7 +6,7 @@ import { useMeta } from 'quasar'
 import { AppName } from '@/constants/global'
 import { getRecordsCountDisplay } from '@/utils/common'
 import { hiddenColumnNames, logColumns } from '@/services/table-columns'
-import { allFields, type AnyDatabaseRecord, type AnyField, type Log } from '@/types/database'
+import { allFields, type AnyDatabaseRecord, type AnyField, type Log } from '@/types/core'
 import useLogger from '@/composables/useLogger'
 import useRoutables from '@/composables/useRoutables'
 import useDialogs from '@/composables/useDialogs'
@@ -42,10 +42,9 @@ onUnmounted(() => {
   subscription?.unsubscribe()
 })
 
-// TODO
 async function onInspect(autoId: number) {
-  const record = await DB.getLog(autoId)
-  inspectDialog('Log', record as AnyDatabaseRecord)
+  const record = (await DB.getLog(Number(autoId))) as AnyDatabaseRecord
+  inspectDialog('Log', record)
 }
 </script>
