+++ src/views/EditView.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { Icon } from '@/types/general'
-import { allFields, type AnyField, type AnyRecord } from '@/types/database'
+import { allFields, type AnyField, type AnyRecord } from '@/types/core'
 import { onMounted, onUnmounted, ref } from 'vue'
 import { extend, uid, useMeta } from 'quasar'
 import { AppName } from '@/constants/global'
